+++ plugins/clipboard-manager/build.rs
@@ -10,7 +10,7 @@ fn main() {
     {
         println!("{error:#}");
         // when building documentation for Android the plugin build result is irrelevant to the crate itself
-        if !(cfg!(feature = "dox") && std::env::var("TARGET").unwrap().contains("android")) {
+        if !(cfg!(docsrs) && std::env::var("TARGET").unwrap().contains("android")) {
             std::process::exit(1);
         }
     }