+++ plugins/notification/android/src/main/java/ChannelManager.kt
@@ -12,21 +12,42 @@ import android.graphics.Color
 import android.media.AudioAttributes
 import android.net.Uri
 import android.os.Build
-import androidx.core.app.NotificationCompat
 import app.tauri.Logger
+import app.tauri.annotation.InvokeArg
 import app.tauri.plugin.Invoke
-import app.tauri.plugin.JSArray
-import app.tauri.plugin.JSObject
+import com.fasterxml.jackson.annotation.JsonValue
 
-private const val CHANNEL_ID = "id"
-private const val CHANNEL_NAME = "name"
-private const val CHANNEL_DESCRIPTION = "description"
-private const val CHANNEL_IMPORTANCE = "importance"
-private const val CHANNEL_VISIBILITY = "visibility"
-private const val CHANNEL_SOUND = "sound"
-private const val CHANNEL_VIBRATE = "vibration"
-private const val CHANNEL_USE_LIGHTS = "lights"
-private const val CHANNEL_LIGHT_COLOR = "lightColor"
+enum class Importance(@JsonValue val value: Int) {
+  None(0),
+  Min(1),
+  Low(2),
+  Default(3),
+  High(4);
+}
+
+enum class Visibility(@JsonValue val value: Int) {
+  Secret(-1),
+  Private(0),
+  Public(1);
+}
+
+@InvokeArg
+class Channel {
+  lateinit var id: String
+  lateinit var name: String
+  var description: String? = null
+  var sound: String? = null
+  var lights: Boolean? = null
+  var lightsColor: String? = null
+  var vibration: Boolean? = null
+  var importance: Importance? = null
+  var visibility: Visibility? = null
+}
+
+@InvokeArg
+class DeleteChannelArgs {
+  lateinit var id: String
+}
 
 class ChannelManager(private var context: Context) {
   private var notificationManager: NotificationManager? = null
@@ -38,32 +59,7 @@ class ChannelManager(private var context: Context) {
 
   fun createChannel(invoke: Invoke) {
     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-      val channel = JSObject()
-      if (invoke.getString(CHANNEL_ID) != null) {
-        channel.put(CHANNEL_ID, invoke.getString(CHANNEL_ID))
-      } else {
-        invoke.reject("Channel missing identifier")
-        return
-      }
-      if (invoke.getString(CHANNEL_NAME) != null) {
-        channel.put(CHANNEL_NAME, invoke.getString(CHANNEL_NAME))
-      } else {
-        invoke.reject("Channel missing name")
-        return
-      }
-      channel.put(
-        CHANNEL_IMPORTANCE,
-        invoke.getInt(CHANNEL_IMPORTANCE, NotificationManager.IMPORTANCE_DEFAULT)
-      )
-      channel.put(CHANNEL_DESCRIPTION, invoke.getString(CHANNEL_DESCRIPTION, ""))
-      channel.put(
-        CHANNEL_VISIBILITY,
-        invoke.getInt(CHANNEL_VISIBILITY, NotificationCompat.VISIBILITY_PUBLIC)
-      )
-      channel.put(CHANNEL_SOUND, invoke.getString(CHANNEL_SOUND))
-      channel.put(CHANNEL_VIBRATE, invoke.getBoolean(CHANNEL_VIBRATE, false))
-      channel.put(CHANNEL_USE_LIGHTS, invoke.getBoolean(CHANNEL_USE_LIGHTS, false))
-      channel.put(CHANNEL_LIGHT_COLOR, invoke.getString(CHANNEL_LIGHT_COLOR))
+      val channel = invoke.parseArgs(Channel::class.java)
       createChannel(channel)
       invoke.resolve()
     } else {
@@ -71,18 +67,18 @@ class ChannelManager(private var context: Context) {
     }
   }
 
-  private fun createChannel(channel: JSObject) {
+  private fun createChannel(channel: Channel) {
     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
       val notificationChannel = NotificationChannel(
-        channel.getString(CHANNEL_ID),
-        channel.getString(CHANNEL_NAME),
-        channel.getInteger(CHANNEL_IMPORTANCE)!!
+        channel.id,
+        channel.name,
+        (channel.importance ?: Importance.Default).value
       )
-      notificationChannel.description = channel.getString(CHANNEL_DESCRIPTION)
-      notificationChannel.lockscreenVisibility = channel.getInteger(CHANNEL_VISIBILITY, android.app.Notification.VISIBILITY_PRIVATE)
-      notificationChannel.enableVibration(channel.getBoolean(CHANNEL_VIBRATE, false))
-      notificationChannel.enableLights(channel.getBoolean(CHANNEL_USE_LIGHTS, false))
-      val lightColor = channel.getString(CHANNEL_LIGHT_COLOR)
+      notificationChannel.description = channel.description
+      notificationChannel.lockscreenVisibility = (channel.visibility ?: Visibility.Private).value
+      notificationChannel.enableVibration(channel.vibration ?: false)
+      notificationChannel.enableLights(channel.lights ?: false)
+      val lightColor = channel.lightsColor ?: ""
       if (lightColor.isNotEmpty()) {
         try {
           notificationChannel.lightColor = Color.parseColor(lightColor)
@@ -94,7 +90,7 @@ class ChannelManager(private var context: Context) {
           )
         }
       }
-      var sound = channel.getString(CHANNEL_SOUND)
+      var sound = channel.sound ?: ""
       if (sound.isNotEmpty()) {
         if (sound.contains(".")) {
           sound = sound.substring(0, sound.lastIndexOf('.'))
@@ -113,8 +109,8 @@ class ChannelManager(private var context: Context) {
 
   fun deleteChannel(invoke: Invoke) {
     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-      val channelId = invoke.getString("id")
-      notificationManager?.deleteNotificationChannel(channelId)
+      val args = invoke.parseArgs(DeleteChannelArgs::class.java)
+      notificationManager?.deleteNotificationChannel(args.id)
       invoke.resolve()
     } else {
       invoke.reject("channel not available")
@@ -125,28 +121,29 @@ class ChannelManager(private var context: Context) {
     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
       val notificationChannels: List<NotificationChannel> =
         notificationManager?.notificationChannels ?: listOf()
-      val channels = JSArray()
+
+      val channels = mutableListOf<Channel>()
+
       for (notificationChannel in notificationChannels) {
-        val channel = JSObject()
-        channel.put(CHANNEL_ID, notificationChannel.id)
-        channel.put(CHANNEL_NAME, notificationChannel.name)
-        channel.put(CHANNEL_DESCRIPTION, notificationChannel.description)
-        channel.put(CHANNEL_IMPORTANCE, notificationChannel.importance)
-        channel.put(CHANNEL_VISIBILITY, notificationChannel.lockscreenVisibility)
-        channel.put(CHANNEL_SOUND, notificationChannel.sound)
-        channel.put(CHANNEL_VIBRATE, notificationChannel.shouldVibrate())
-        channel.put(CHANNEL_USE_LIGHTS, notificationChannel.shouldShowLights())
-        channel.put(
-          CHANNEL_LIGHT_COLOR, String.format(
-            "#%06X",
-            0xFFFFFF and notificationChannel.lightColor
-          )
+        val channel = Channel()
+        channel.id = notificationChannel.id
+        channel.name = notificationChannel.name.toString()
+        channel.description = notificationChannel.description
+        channel.sound = notificationChannel.sound.toString()
+        channel.lights = notificationChannel.shouldShowLights()
+        String.format(
+          "#%06X",
+          0xFFFFFF and notificationChannel.lightColor
         )
-        channels.put(channel)
+        channel.vibration = notificationChannel.shouldVibrate()
+        channel.importance = Importance.values().firstOrNull { it.value == notificationChannel.importance }
+        channel.visibility = Visibility.values().firstOrNull { it.value == notificationChannel.lockscreenVisibility }
+
+        channels.add(channel)
       }
-      val result = JSObject()
-      result.put("channels", channels)
-      invoke.resolve(result)
+
+      invoke.resolveObject(channels)
+
     } else {
       invoke.reject("channel not available")
     }