+++ plugins/notification/android/src/main/java/NotificationStorage.kt
@@ -7,7 +7,10 @@ package app.tauri.notification
 import android.content.Context
 import android.content.SharedPreferences
 import app.tauri.plugin.JSObject
+import com.fasterxml.jackson.databind.ObjectMapper
+import com.fasterxml.jackson.module.kotlin.registerKotlinModule
 import org.json.JSONException
+import java.lang.Exception
 import java.text.ParseException
 
 // Key for private preferences
@@ -22,7 +25,7 @@ class NotificationStorage(private val context: Context) {
     for (request in localNotifications) {
       if (request.isScheduled) {
         val key: String = request.id.toString()
-        editor.putString(key, request.source.toString())
+        editor.putString(key, request.sourceJson.toString())
       }
     }
     editor.apply()
@@ -43,57 +46,29 @@ class NotificationStorage(private val context: Context) {
       val notifications = ArrayList<Notification>()
       for (key in all.keys) {
         val notificationString = all[key] as String?
-        val jsNotification = getNotificationFromJSONString(notificationString)
-        if (jsNotification != null) {
-          try {
-            val notification =
-              Notification.fromJSObject(jsNotification)
-            notifications.add(notification)
-          } catch (_: ParseException) {
-          }
-        }
+        try {
+          val notification = ObjectMapper().registerKotlinModule().readValue(notificationString, Notification::class.java)
+          notifications.add(notification)
+        } catch (_: Exception) { }
       }
       return notifications
     }
     return ArrayList()
   }
 
-  private fun getNotificationFromJSONString(notificationString: String?): JSObject? {
-    if (notificationString == null) {
-      return null
-    }
-    val jsNotification = try {
-      JSObject(notificationString)
-    } catch (ex: JSONException) {
-      return null
-    }
-    return jsNotification
-  }
-
-  fun getSavedNotificationAsJSObject(key: String?): JSObject? {
+  fun getSavedNotification(key: String): Notification? {
     val storage = getStorage(NOTIFICATION_STORE_ID)
     val notificationString = try {
       storage.getString(key, null)
     } catch (ex: ClassCastException) {
       return null
     } ?: return null
-    
-    val jsNotification = try {
-      JSObject(notificationString)
-    } catch (ex: JSONException) {
-      return null
-    }
-    return jsNotification
-  }
 
-  fun getSavedNotification(key: String?): Notification? {
-    val jsNotification = getSavedNotificationAsJSObject(key) ?: return null
-    val notification = try {
-      Notification.fromJSObject(jsNotification)
-    } catch (ex: ParseException) {
-      return null
+    return try {
+      ObjectMapper().registerKotlinModule().readValue(notificationString, Notification::class.java)
+    } catch (ex: JSONException) {
+      null
     }
-    return notification
   }
 
   fun deleteNotification(id: String?) {
@@ -106,15 +81,16 @@ class NotificationStorage(private val context: Context) {
     return context.getSharedPreferences(key, Context.MODE_PRIVATE)
   }
 
-  fun writeActionGroup(typesMap: Map<String, List<NotificationAction>>) {
-    for ((id, notificationActions) in typesMap) {
-      val editor = getStorage(ACTION_TYPES_ID + id).edit()
+  fun writeActionGroup(actions: List<ActionType>) {
+    for (type in actions) {
+      val i = type.id
+      val editor = getStorage(ACTION_TYPES_ID + type.id).edit()
       editor.clear()
-      editor.putInt("count", notificationActions.size)
-      for (i in notificationActions.indices) {
-        editor.putString("id$i", notificationActions[i].id)
-        editor.putString("title$i", notificationActions[i].title)
-        editor.putBoolean("input$i", notificationActions[i].input)
+      editor.putInt("count", type.actions.size)
+      for (action in type.actions) {
+        editor.putString("id$i", action.id)
+        editor.putString("title$i", action.title)
+        editor.putBoolean("input$i", action.input ?: false)
       }
       editor.apply()
     }
@@ -128,7 +104,12 @@ class NotificationStorage(private val context: Context) {
       val id = storage.getString("id$i", "")
       val title = storage.getString("title$i", "")
       val input = storage.getBoolean("input$i", false)
-      actions[i] = NotificationAction(id, title, input)
+
+      val action = NotificationAction()
+      action.id = id ?: ""
+      action.title = title
+      action.input = input
+      actions[i] = action
     }
     return actions
   }